requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
chardet==3.0.4
Django==2.1.2
djangorestframework==3.9.0
pdfminer.six==20170720
Pillow==5.3.0
psycopg2==2.7.5
pycryptodome==3.6.6
python-dotenv==0.9.1
pytz==2018.5
six==1.11.0
